Model and Endpoint Versioning Checks
Update: enhanced logic around model registering and endpoint deployment to check for existing model and version and not upload or deploy in duplicate.  Also creates model and/or endpoint when they are missing.  This makes the notebook stand-alone within the 05 series - not even requiring 05.  Rerunning a notebook will also create new versions of a model while also deploying the new version to the 05_fraud endpoint.